+++ resources/lang/cs/harvest.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'status_message_progress' => 'Zpracováva se (%d/%d)',
+    'status_message_completed' =>
+        'Zpracovaných záznamů: %d:' . PHP_EOL .
+        '%d nových' . PHP_EOL .
+        '%d upravených' . PHP_EOL .
+        '%d smazaných' . PHP_EOL .
+        '%d přeskočených.' . PHP_EOL .
+        'Trvalo to %.0fs',
+];